Add URL shortening capability with YAML and Map handlers
Introduced a URL shortening functionality in main.go which involves using a Map handler and a YAML handler. The map handler takes in a mapping of path strings to their corresponding URLs as well as a fallback http.Handler. The YAML handler, on the other hand, parses provided YAML to map any paths to the corresponding URLs, with its own fallback http.Handler. Additional methods were also created for unmarshalling YAML data and creating a YAML map. A .gitignore file, go.mod and go.sum files were added as well.